update chicago bike demo
