--- docs/source/examples/demo/chicago_bike/future.csv
@@ -1,56 +1,0 @@
-trip_id,gender,starttime,stoptime,tripduration,temperature,events,from_station_id,dpcapacity_start,to_station_id,dpcapacity_end
-16734175,Male,2017-10-01 00:48:00,2017-10-01 00:55:00,7.416666666666668,53.1,cloudy,288,11.0,60,19.0
-16734170,Female,2017-10-01 00:46:00,2017-10-01 00:56:00,9.2,53.1,cloudy,289,19.0,340,15.0
-16734169,Male,2017-10-01 00:46:00,2017-10-01 01:00:00,13.516666666666667,53.1,cloudy,256,15.0,461,15.0
-16734168,Male,2017-10-01 00:46:00,2017-10-01 00:56:00,9.9,53.1,cloudy,289,19.0,340,15.0
-16734165,Male,2017-10-01 00:45:00,2017-10-01 01:05:00,19.866666666666667,53.1,cloudy,110,23.0,21,15.0
-16734164,Male,2017-10-01 00:45:00,2017-10-01 01:05:00,19.866666666666667,53.1,cloudy,110,23.0,21,15.0
-16734163,Male,2017-10-01 00:45:00,2017-10-01 00:59:00,14.4,53.1,cloudy,291,19.0,142,31.0
-16734162,Male,2017-10-01 00:44:00,2017-10-01 00:49:00,4.766666666666667,53.1,cloudy,299,15.0,117,23.0
-16734161,Female,2017-10-01 00:43:00,2017-10-01 00:47:00,4.05,53.1,cloudy,302,19.0,327,19.0
-16734160,Female,2017-10-01 00:43:00,2017-10-01 00:57:00,13.85,53.1,cloudy,302,19.0,236,15.0
-16734159,Male,2017-10-01 00:43:00,2017-10-01 00:45:00,2.1666666666666665,53.1,cloudy,227,23.0,229,19.0
-16734156,Male,2017-10-01 00:41:00,2017-10-01 01:10:00,28.98333333333333,53.1,cloudy,112,15.0,159,11.0
-16734151,Female,2017-10-01 00:38:00,2017-10-01 00:48:00,9.85,53.1,cloudy,69,19.0,350,15.0
-16734147,Male,2017-10-01 00:36:00,2017-10-01 01:01:00,25.38333333333333,53.1,cloudy,117,23.0,302,19.0
-16734146,Male,2017-10-01 00:36:00,2017-10-01 00:50:00,14.183333333333335,53.1,cloudy,94,31.0,131,15.0
-16734144,Male,2017-10-01 00:32:00,2017-10-01 00:39:00,6.4833333333333325,53.1,cloudy,117,23.0,230,19.0
-16734143,Male,2017-10-01 00:32:00,2017-10-01 01:00:00,27.93333333333333,53.1,cloudy,237,15.0,90,47.0
-16734142,Male,2017-10-01 00:32:00,2017-10-01 01:00:00,27.9,53.1,cloudy,237,15.0,90,47.0
-16734141,Male,2017-10-01 00:31:00,2017-10-01 00:49:00,18.016666666666666,53.1,cloudy,488,11.0,495,15.0
-16734140,Female,2017-10-01 00:31:00,2017-10-01 00:49:00,18.35,53.1,cloudy,488,11.0,495,15.0
-16734138,Male,2017-10-01 00:30:00,2017-10-01 00:39:00,8.733333333333333,53.1,cloudy,117,23.0,227,23.0
-16734136,Male,2017-10-01 00:29:00,2017-10-01 00:55:00,25.85,53.1,cloudy,109,19.0,109,19.0
-16734135,Male,2017-10-01 00:29:00,2017-10-01 00:50:00,21.63333333333333,53.1,cloudy,229,19.0,453,15.0
-16734134,Male,2017-10-01 00:28:00,2017-10-01 00:31:00,3.2,53.1,cloudy,112,15.0,77,31.0
-16734133,Male,2017-10-01 00:28:00,2017-10-01 00:41:00,13.466666666666667,53.1,cloudy,74,23.0,213,15.0
-16734132,Female,2017-10-01 00:27:00,2017-10-01 00:41:00,13.983333333333333,53.1,cloudy,74,23.0,213,15.0
-16734129,Male,2017-10-01 00:25:00,2017-10-01 00:33:00,8.616666666666667,53.1,cloudy,69,19.0,219,19.0
-16734122,Male,2017-10-01 00:22:00,2017-10-01 00:29:00,6.666666666666668,53.1,cloudy,94,31.0,349,15.0
-16734119,Male,2017-10-01 00:21:00,2017-10-01 00:42:00,20.33333333333333,53.1,cloudy,51,39.0,40,15.0
-16734118,Female,2017-10-01 00:21:00,2017-10-01 00:33:00,12.433333333333335,53.1,cloudy,133,31.0,291,19.0
-16734117,Female,2017-10-01 00:21:00,2017-10-01 00:33:00,12.5,53.1,cloudy,133,31.0,291,19.0
-16734116,Female,2017-10-01 00:21:00,2017-10-01 00:33:00,12.566666666666665,53.1,cloudy,133,31.0,291,19.0
-16734111,Male,2017-10-01 00:17:00,2017-10-01 00:23:00,5.9,53.1,cloudy,299,15.0,349,15.0
-16734110,Male,2017-10-01 00:17:00,2017-10-01 00:26:00,9.566666666666666,53.1,cloudy,112,15.0,43,43.0
-16734105,Male,2017-10-01 00:15:00,2017-10-01 00:21:00,5.066666666666666,53.1,cloudy,256,15.0,306,19.0
-16734104,Female,2017-10-01 00:15:00,2017-10-01 00:23:00,8.366666666666667,53.1,cloudy,227,23.0,234,19.0
-16734102,Male,2017-10-01 00:14:00,2017-10-01 00:21:00,7.183333333333334,53.1,cloudy,112,15.0,84,19.0
-16734101,Male,2017-10-01 00:11:00,2017-10-01 00:21:00,9.95,53.1,cloudy,51,39.0,51,39.0
-16734094,Male,2017-10-01 00:08:00,2017-10-01 00:16:00,8.016666666666667,53.1,cloudy,109,19.0,19,16.0
-16734093,Female,2017-10-01 00:08:00,2017-10-01 00:16:00,8.316666666666666,53.1,cloudy,109,19.0,136,15.0
-16734092,Male,2017-10-01 00:08:00,2017-10-01 00:27:00,19.15,53.1,cloudy,153,19.0,138,15.0
-16734091,Female,2017-10-01 00:08:00,2017-10-01 00:27:00,19.533333333333328,53.1,cloudy,153,19.0,138,15.0
-16734090,Male,2017-10-01 00:07:00,2017-10-01 00:13:00,5.583333333333332,53.1,cloudy,87,19.0,144,19.0
-16734089,Male,2017-10-01 00:07:00,2017-10-01 00:13:00,5.833333333333332,53.1,cloudy,87,19.0,144,19.0
-16734088,Female,2017-10-01 00:07:00,2017-10-01 00:28:00,21.516666666666666,53.1,cloudy,488,11.0,488,11.0
-16734087,Male,2017-10-01 00:07:00,2017-10-01 00:28:00,21.516666666666666,53.1,cloudy,488,11.0,488,11.0
-16734084,Female,2017-10-01 00:06:00,2017-10-01 00:12:00,5.55,53.1,cloudy,62,27.0,273,23.0
-16734083,Male,2017-10-01 00:06:00,2017-10-01 00:21:00,15.166666666666664,53.1,cloudy,62,27.0,273,23.0
-16734081,Male,2017-10-01 00:05:00,2017-10-01 00:21:00,16.7,53.1,cloudy,229,19.0,144,19.0
-16734080,Male,2017-10-01 00:04:00,2017-10-01 00:12:00,7.383333333333334,53.1,cloudy,291,19.0,60,19.0
-16734075,Male,2017-10-01 00:04:00,2017-10-01 00:10:00,6.0,53.1,cloudy,229,19.0,330,19.0
-16734072,Female,2017-10-01 00:01:00,2017-10-01 00:12:00,11.066666666666665,53.1,cloudy,220,23.0,87,19.0
-16734071,Male,2017-10-01 00:01:00,2017-10-01 00:12:00,11.033333333333333,53.1,cloudy,220,23.0,87,19.0
-16734070,Male,2017-10-01 00:01:00,2017-10-01 00:15:00,13.95,53.1,cloudy,289,19.0,226,15.0
-16734067,Female,2017-10-01 00:00:00,2017-10-01 00:06:00,6.016666666666668,53.1,cloudy,288,11.0,289,19.0
